+++ lib/asset-definitions/babel.js
