+++ lib/asset-definitions/coffeescript.js
