+++ lib/asset-definitions/ejs.js
