+++ lib/asset-definitions/index.js
@@ -13,10 +13,10 @@ var html = {
       extensions: ['ejs'],
       render: require('./ejs'),
     },
-    // jade: {
-    //   extensions: ['jade'],
-    //   render: require('./jade'),
-    // },
+    jade: {
+      extensions: ['jade'],
+      render: require('./jade'),
+    },
   }
 };
