+++ lib/asset-definitions/jade.js
@@ -1,0 +1,10 @@
+'use strict';
+
+var resolveModule = require('../util/Module').resolveModule;
+
+module.exports = function renderJade(str) {
+  return resolveModule('jade')
+    .then(function (jade) {
+      return jade.render(str, {pretty: true});
+    });
+};
