+++ lib/asset-definitions/stylus.js
