+++ lib/asset-definitions/typescript.js
