+++ lib/preStart.js
@@ -4,6 +4,7 @@ var assert = require('assert');
 var chalk = require('chalk');
 var Observable = require('rx').Observable;
 var R = require('ramda');
+var Assets = require('./util/Assets');
 var Functional = require('./util/Functional');
 var Log = require('../lib/util/Log');
 var File = require('./util/File');
@@ -37,24 +38,25 @@ module.exports = function preStart(start) {
       var name = results[1];
       return Playground.parse(Playground.getExt(name), content);
     })
-    .flatMap(function (playground) {
-      var htmlObs = Playground.createAssetFile(playground, 'html', function (ext) {
-        hasCreatedFromTmpl = true;
-        return ext === 'html' ? File.readFileObs(CONST.HTML_EXAMPLE) : Observable.just('');
-      });
-
-      var jsObs = Playground.createAssetFile(playground, 'js', function (ext) {
-        hasCreatedFromTmpl = true;
-        return Observable.just('');
-      });
-
-      var cssObs = Playground.createAssetFile(playground, 'css', function (ext) {
-        hasCreatedFromTmpl = true;
-        return Observable.just('');
-      });
-
-      return Observable.combineLatest(htmlObs, jsObs, cssObs);
-    })
+    .flatMap(R.pipe(
+      R.repeat(R.__, 3),
+      R.evolve([
+        Assets.createAssetFile('html', function (ext) {
+          hasCreatedFromTmpl = true;
+          return ext === 'html' ? File.readFileObs(CONST.HTML_EXAMPLE) : Observable.just('');
+        }),
+        Assets.createAssetFile('js', function (ext) {
+          hasCreatedFromTmpl = true;
+          return Observable.just('');
+        }),
+        Assets.createAssetFile('css', function (ext) {
+          hasCreatedFromTmpl = true;
+          return Observable.just('');
+        })
+      ]),
+      R.values,
+      Observable.combineLatest
+    ))
     .subscribeOnCompleted(function () {
       if (hasCreatedFromTmpl) {
         Log.info('Run ' + chalk.green('wpg') + ' again to start live-reload server');
